# openapi.yaml
openapi: 3.1.0
info:
  title: Jitsi Portal API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/users/search:
    get:
      summary: Search users by name and organization within tenant
      operationId: searchUsers
      security:
        - oauth2: []
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
            minLength: 2
        - name: organization
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Search results (max 50)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfileSummary'
        '403':
          description: Tenant access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
components:
  schemas:
...
    ProblemDetailResponse:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        properties:
          type: object
          properties:
            errorCode:
              type: string
            traceId:
              type: string
          required:
            - errorCode
            - traceId
...
    UserProfileSummary:
      type: object
      properties:
        subjectId:
          type: string
        fullName:
          type: string
        organization:
          type: string
        position:
          type: string
      required:
        - subjectId
        - fullName
        - organization
        - position